requirements.txt
celery
channels
channels-redis
crispy-bootstrap5
daphne
Django
django-auth-ldap
django-celery-results
django-crispy-forms
django-oauth2-authcodeflow
django-timezone-field
djangorestframework
drf-spectacular
ocpp
Pillow
psycopg2-binary
redis
requests
sanic
websockets
django-countries
Twisted[tls,http2]
elasticsearch
djangorestframework-simplejwt[crypto]
django-filter
